adding pkg build
