--- ailinter/config.yaml
@@ -1,2 +1,0 @@
-RULE_GUIDE: "human_guide_1.md"
-DEFAULT_TEMPERATURE: 0.2
